das-schcanner
This is das-schcanner, a tool used to scan a network range you give it and perform different types of scans again it. Scans performed are arp-scan, fping and nbtscan. This then creates 3 results .txt files from them 3 scans, then goes and greps through all 3 files and removes in duplicates and saves it all in a final .txt file called ALLDONE.txt.